Switch personal data and edit fix
