+++ src/main/java/lv/ctco/javaschool/auth/entity/domain/User.java
@@ -1,12 +1,6 @@
 package lv.ctco.javaschool.auth.entity.domain;
 
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.EnumType;
-import javax.persistence.Enumerated;
-import javax.persistence.GeneratedValue;
-import javax.persistence.Id;
-import javax.persistence.Table;
+import javax.persistence.*;
 
 @Entity
 @Table(name = "users")
