+++ src/main/java/lv/ctco/javaschool/goal/entity/domain/Goal.java
@@ -2,14 +2,7 @@ package lv.ctco.javaschool.goal.entity.domain;
 
 import lv.ctco.javaschool.auth.entity.domain.User;
 
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.Id;
-import javax.persistence.JoinColumn;
-import javax.persistence.JoinTable;
-import javax.persistence.ManyToOne;
-import javax.persistence.OneToMany;
-import javax.persistence.Table;
+import javax.persistence.*;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.Set;
