+++ src/main/java/lv/ctco/javaschool/goal/entity/domain/Tag.java
@@ -1,10 +1,6 @@
 package lv.ctco.javaschool.goal.entity.domain;
 
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.Id;
-import javax.persistence.Table;
+import javax.persistence.*;
 
 @Entity
 @Table(name = "tags")
