+++ src/main/webapp/app/js/start-page.js
@@ -1,15 +1,5 @@
 var path = "";
 
-function switchPersonalData() {
-    var checkbox = document.getElementById("personal-data");
-    var block = document.getElementById("personal-block");
-    if (checkbox.checked) {
-        block.classList.remove("w3-hide");
-    } else {
-        block.classList.add("w3-hide");
-    }
-}
-
 function showUserProfile() {
     fetch(path + "/api/auth/myprofile", {
         "method": "GET",
